--- book-service/src/main/java/com/beta/bookservice/model/Phone.java
@@ -1,21 +1,0 @@
-package com.beta.bookservice.model;
-
-public class Phone {
-
-    private String phone;
-
-    public Phone() {
-    }
-
-    public Phone(String phone) {
-        this.phone = phone;
-    }
-
-    public String getPhone() {
-        return phone;
-    }
-
-    public void setPhone(String phone) {
-        this.phone = phone;
-    }
-}
